src/requirements.txt
selenium
tk
setuptools
reportlab
PyPDF2
python-dateutil
pyautogui